+++ screenlockpin.koplugin/_meta.lua
@@ -6,8 +6,10 @@ return {
     fullname = _("ScreenLock PIN"),
     description = _([[Protect your device privacy with a PIN.]]),
 
-    -- additional meta information; no effect in code
+    -- used to check for latest plugin updates
+    update_url = "https://api.github.com/repos/oleasteo/koreader-screenlockpin/releases/latest",
     version = "2025.11-1",
+    -- additional meta information; no effect in code
     website_url = "https://github.com/oleasteo/koreader-screenlockpin",
     repository_url = "https://github.com/oleasteo/koreader-screenlockpin.git",
 }